Langchain/URL-Website-Content-Summarizer/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
streamlit
langchain-groq
langchain-chroma
langchain-huggingface
sentence-transformers
langchain-cohere
youtube-transcript-api
pytube
validators

